fix warning

+++ fem/DGIntegrators.cpp
@@ -318,7 +318,7 @@ namespace ngfem
       ELEMENT_TYPE eltype2 = volumefel2.ElementType();
       eltype2 = volumefel2.ElementType();
       int nd2 = fel2_l2->GetNDof();
-      double maxorder = max(fel1_l2->Order(),fel2_l2->Order());
+      int maxorder = max(fel1_l2->Order(),fel2_l2->Order());
       
       elmat = 0.0;
 
@@ -552,7 +552,7 @@ namespace ngfem
       // ELEMENT_TYPE eltype2 = eltype1;
       int nd2 = 0;
 
-      double maxorder = fel1_l2->Order();
+      int maxorder = fel1_l2->Order();
 
       
       elmat = 0.0;
@@ -616,13 +616,13 @@ namespace ngfem
 	    case DG_FORMULATIONS::NIPG:
 	      dmat(0,1) = 1; 
 // 	      dmat(1,1) = alpha * sqr (maxorder) * (len1/det1);
-	      dmat(1,1) = alpha * ((maxorder+1)*(maxorder+D)/D * len1) *(1.0/det1);
+	      dmat(1,1) = alpha * ((maxorder+1.0)*(maxorder+D)/D * len1) *(1.0/det1);
 	      break;
 	    case DG_FORMULATIONS::IP:
 	    default:
 	      dmat(0,1) = -1; 
 // 	      dmat(1,1) = alpha * sqr (maxorder) * (len1/det1);
-	      dmat(1,1) = alpha * ((maxorder+1)*(maxorder+D)/D * len1) *(1.0/det1);
+	      dmat(1,1) = alpha * ((maxorder+1.0)*(maxorder+D)/D * len1) *(1.0/det1);
 	      break;	      
 	  }
 	  dmat *= lam * len1 * ir_facet[l].Weight();
@@ -696,7 +696,7 @@ namespace ngfem
       for (int i=0; i<D; i++){
 	normal_ref1(i) = normals1[LocalFacetNr][i];
       }
-      double maxorder = fel1_l2->Order();
+      int maxorder = fel1_l2->Order();
       const IntegrationRule & ir_facet =
 	SelectIntegrationRule (etfacet, 2*maxorder);
       if (maxorder==0) maxorder=1;
@@ -734,13 +734,13 @@ namespace ngfem
 	    case DG_FORMULATIONS::NIPG:
 	      dvec(0) = 1; 
 // 	      dvec(1) = alpha * sqr (maxorder) * (len1/det1);
-	      dvec(1) = alpha * ((maxorder+1)*(maxorder+D)/D * len1) *(1.0/det1);	      
+	      dvec(1) = alpha * ((maxorder+1.0)*(maxorder+D)/D * len1) *(1.0/det1);	      
 	      break;
 	    case DG_FORMULATIONS::IP:
 	    default:
 	      dvec(0) = -1;
 // 	      dvec(1) = alpha * sqr (maxorder) * (len1/det1);
-	      dvec(1) = alpha * ((maxorder+1)*(maxorder+D)/D * len1) *(1.0/det1);	      
+	      dvec(1) = alpha * ((maxorder+1.0)*(maxorder+D)/D * len1) *(1.0/det1);	      
 	      break;	      
 	  }	  
 	  
@@ -810,7 +810,7 @@ namespace ngfem
       for (int i=0; i<D; i++){
 	normal_ref1(i) = normals1[LocalFacetNr][i];
       }
-      double maxorder = fel1_l2->Order();
+      int maxorder = fel1_l2->Order();
       const IntegrationRule & ir_facet =
 	SelectIntegrationRule (etfacet, 2*maxorder);
       if (maxorder==0) maxorder=1;
